Add list of access grants on user page
List contains explicit access rules that have been created for this
user, to server accounts or groups. Access rules that apply through the
user's group memberships are not included (group memberships are
already listed on the page and can be followed for this information),
nor is automatic access to servers with LDAP access included.